converts files to ts
