--- src/App.js
@@ -1,10 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-
-function App() {
-  return (
-    
-  );
-}
-
-export default App;
